requirements.txt
aiogram==3.0.0b5
aiohttp==3.8.1
python-dotenv==1.0.0